requirements.txt
tensorflow>=2.12.0
numpy>=1.23.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.2.0
opencv-python>=4.7.0
streamlit>=1.28.0
Pillow>=9.5.0
tf-keras>=2.16.0
